+++ realm/realm-library/src/main/java/io/realm/log/RealmLog.java
@@ -18,9 +18,6 @@ package io.realm.log;
 
 import android.util.Log;
 
-import java.util.IdentityHashMap;
-import java.util.Map;
-
 /**
  * Global logger used by all Realm components.
  * Custom loggers can be added by registering classes implementing {@link RealmLogger}.
@@ -29,56 +26,6 @@ public final class RealmLog {
 
     @SuppressWarnings("FieldCanBeLocal")
     private static String REALM_JAVA_TAG = "REALM_JAVA";
-
-    /**
-     * To convert the old {@link Logger} to the new {@link RealmLogger}.
-     */
-    private static class LoggerAdapter implements RealmLogger {
-        private Logger logger;
-        private static final Map<Logger, LoggerAdapter> loggerMap = new IdentityHashMap<Logger, LoggerAdapter>();
-
-        LoggerAdapter(Logger logger) {
-            this.logger = logger;
-            if (loggerMap.containsKey(logger)) {
-                throw new IllegalStateException(String.format("Logger %s exists in the map!", logger.toString()));
-            }
-            loggerMap.put(logger, this);
-        }
-
-        static RealmLogger removeLogger(Logger logger) {
-            return loggerMap.remove(logger);
-        }
-
-        static void clear() {
-            loggerMap.clear();
-        }
-
-        @Override
-        public void log(int level, String tag, Throwable throwable, String message) {
-            switch (level) {
-                case LogLevel.TRACE:
-                    logger.trace(throwable, message);
-                    break;
-                case LogLevel.INFO:
-                    logger.info(throwable, message);
-                    break;
-                case LogLevel.DEBUG:
-                    logger.debug(throwable, message);
-                    break;
-                case LogLevel.WARN:
-                    logger.warn(throwable, message);
-                    break;
-                case LogLevel.ERROR:
-                    logger.error(throwable, message);
-                    break;
-                case LogLevel.FATAL:
-                    logger.fatal(throwable, message);
-                    break;
-                default:
-                    throw new IllegalArgumentException("Level: " + level + " cannot be logged.");
-            }
-        }
-    }
 
     /**
      * Adds a logger implementation that will be notified on log events.
@@ -90,18 +37,6 @@ public final class RealmLog {
             throw new IllegalArgumentException("A non-null logger has to be provided");
         }
         nativeAddLogger(logger);
-    }
-
-    /**
-     * Adds a logger implementation that will be notified on log events.
-     *
-     * @param logger the reference to a {@link Logger} implementation.
-     * @deprecated use {@link #add(RealmLogger)} instead.
-     */
-    public static void add(Logger logger) {
-        synchronized (LoggerAdapter.class) {
-            add(new LoggerAdapter(logger));
-        }
     }
 
     /**
@@ -136,33 +71,11 @@ public final class RealmLog {
     }
 
     /**
-     * Removes the given logger if it is currently added.
-     *
-     * @return {@code true} if the logger was removed, {@code false} otherwise.
-     * @deprecated use {@link #remove(RealmLogger)} instead.
-     */
-    public static boolean remove(Logger logger) {
-        synchronized (LoggerAdapter.class) {
-            if (logger == null) {
-                throw new IllegalArgumentException("A non-null logger has to be provided");
-            }
-            RealmLogger adaptor = LoggerAdapter.removeLogger(logger);
-            if (adaptor != null) {
-                nativeRemoveLogger(adaptor);
-            }
-        }
-        return true;
-    }
-
-    /**
      * Removes all loggers. The default native logger will be removed as well. Use {@link #registerDefaultLogger()} to
      * add it back.
      */
     public static void clear() {
-        synchronized (LoggerAdapter.class) {
-            nativeClearLoggers();
-            LoggerAdapter.clear();
-        }
+        nativeClearLoggers();
     }
 
     /**